# swagger.yml
openapi: 3.0.0
paths:
...
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Authenticate and issue access/refresh tokens
      tags: &ref_0
        - auth
  /api/auth/refresh-token:
    post:
      operationId: AuthController_refreshToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Rotate refresh token and issue a new token pair
      tags: *ref_0
  /api/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponseDto'
      security:
        - bearer: []
      summary: Get current user profile with accessible modules
      tags: *ref_0
  /api/users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      security: &ref_1
        - bearer: []
      summary: Create a user account
      tags: &ref_2
        - users
    get:
      operationId: UsersController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
      responses:
        '200':
          description: ''
      security: *ref_1
      summary: List users with pagination
      tags: *ref_2
  /api/users/{id}:
    get:
      operationId: UsersController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      security: *ref_1
      summary: Get user by ID
      tags: *ref_2
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      security: *ref_1
      summary: Update a user
      tags: *ref_2
    delete:
      operationId: UsersController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security: *ref_1
      summary: Delete a user
      tags: *ref_2
  /api/users/{userId}/roles:
    patch:
      operationId: UsersController_assignRole
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      security: *ref_1
      summary: Assign role to user
      tags: *ref_2
  /api/roles:
    post:
      operationId: RolesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponseDto'
      security: &ref_3
        - bearer: []
      summary: Create a role
      tags: &ref_4
        - roles
    get:
      operationId: RolesController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponseDto'
      security: *ref_3
      summary: List all roles
      tags: *ref_4
  /api/roles/{id}:
    get:
      operationId: RolesController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponseDto'
      security: *ref_3
      summary: Get a role by ID
      tags: *ref_4
    put:
      operationId: RolesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponseDto'
      security: *ref_3
      summary: Update a role
      tags: *ref_4
    delete:
      operationId: RolesController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security: *ref_3
      summary: Delete a role
      tags: *ref_4
...
  /api/website/auth/login:
    post:
      operationId: WebsiteAuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Login for website users
      tags: *ref_46
...
info:
  title: RxSoft Backend API
  description: Phase 3 identity module and API contracts
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: P@ssw0rd!
      required:
        - username
        - password
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresIn:
          type: number
        refreshTokenExpiresIn:
          type: number
      required:
        - accessToken
        - refreshToken
        - accessTokenExpiresIn
        - refreshTokenExpiresIn
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    ModuleInfoDto:
      type: object
      properties:
        id:
          type: string
          example: rxsoft
        name:
          type: string
          example: RxSoft
        description:
          type: string
          example: Pharmacy Admin
        root:
          type: string
          example: /items
      required:
        - id
        - name
        - description
        - root
    MeResponseDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: array
        permissions:
          type: array
          items:
            type: array
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleInfoDto'
      required:
        - id
        - username
        - roles
        - permissions
        - modules
    CreateUserDto:
      type: object
      properties:
        username:
          type: string
          example: cashier01
        password:
          type: string
          example: P@ssw0rd!
        phone:
          type: string
          example: '+1234567890'
        roleCodes:
          example:
            - cashier
          type: array
          items:
            type: string
      required:
        - username
        - password
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: array
      required:
        - id
        - username
        - roles
    Object:
      type: object
      properties: {}
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
          example: cashier01
        password:
          type: string
          example: N3wP@ss!
        roleCodes:
          example:
            - cashier
          type: array
          items:
            type: string
        phone:
          type: string
          example: '+1234567890'
        isActive:
          type: boolean
          example: true
    AssignRoleDto:
      type: object
      properties:
        roleCode:
          type: string
          example: cashier
      required:
        - roleCode
    CreateRoleDto:
      type: object
      properties:
        code:
          type: string
          example: manager
        name:
          type: string
          example: Manager
        description:
          type: string
          example: Can manage products and inventory
        permissionCodes:
          example:
            - products:read
            - inventory:write
          type: array
          items:
            type: string
      required:
        - code
        - name
    RoleResponseDto:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: object
          nullable: true
        permissionCodes:
          type: array
          items:
            type: array
      required:
        - id
        - code
        - name
        - description
        - permissionCodes
    UpdateRoleDto:
      type: object
      properties:
        code:
          type: string
          example: manager
        name:
          type: string
          example: Manager
        description:
          type: string
          example: Can manage products and inventory
        permissionCodes:
          example:
            - products:read
            - products:write
          type: array
          items:
            type: string
...